move ItemPage into items directory
